refactor(domain): split downloader into domain modules and remove re-export shim
- extract platform, artifact, and download models into src/domain
- move target detection tests into domain/download
- remove app_downloader compatibility module and direct imports to domain
- simplify main module declarations for new structure
- update refactor plan and bug audit progress
