+++ src/domain/mod.rs
@@ -1,0 +1,3 @@
+pub mod artifact;
+pub mod download;
+pub mod platform;
