+++ src/main.rs
@@ -11,7 +11,7 @@ use std::net::SocketAddr;
 use tower_http::cors::CorsLayer;
 use utoipa::OpenApi;
 use utoipa_swagger_ui::SwaggerUi;
-mod app_downloader;
+mod domain;
 mod error;
 mod gh;
 mod static_site;
@@ -21,9 +21,10 @@ mod types;
 mod services;
 
 use crate::error::AppError;
+use crate::domain::platform::{TargetArch, TargetOs};
 use crate::services::installer;
 use crate::templates::TEMPLATES;
-use crate::types::{InstallMethod, InstallQueryOptions, ScriptResponse, TargetArch, TargetOs};
+use crate::types::{InstallMethod, InstallQueryOptions, ScriptResponse};
 use log::{debug, info, warn};
 use std::env;
 use std::path::PathBuf;
